chore(BaseInput): refactor and restructure
