--- packages/blade/src/components/Input/BaseInput/BaseInput.d.ts
@@ -1,1 +1,0 @@
-export * from './BaseInput.web';
