+++ packages/blade/src/components/Input/BaseInput/BaseInput.stories.tsx
@@ -2,7 +2,7 @@ import type { ComponentStory, Meta } from '@storybook/react';
 import { Title, Subtitle, Primary, ArgsTable, Stories, PRIMARY_STORY } from '@storybook/addon-docs';
 import { Highlight } from '@storybook/design-system';
 import React from 'react';
-import type { BaseInputProps } from './baseInputHelpers';
+import type { BaseInputProps } from './BaseInput';
 import { BaseInput as BaseInputComponent } from './BaseInput';
 
 export default {
