+++ packages/blade/src/components/Input/BaseInput/StyledBaseInput.d.ts
@@ -1,0 +1,17 @@
+import type { BaseInputProps } from './BaseInput';
+
+export type StyledBaseInputProps = {
+  handleOnChange?: HandleOnChange;
+} & Pick<
+  BaseInputProps,
+  | 'type'
+  | 'name'
+  | 'placeholder'
+  | 'defaultValue'
+  | 'value'
+  | 'isDisabled'
+  | 'isRequired'
+  | 'validationState'
+>;
+
+export { StyledBaseInput } from './StyledBaseInput.web';
